# openapi.yaml
openapi: 3.1.0
info:
  title: Kitsu DeFi API
  description: |
    Gamified DeFi savings app on TON blockchain.
    
    ## Authentication
    Most endpoints require JWT authentication. Get a token by connecting your wallet via `/auth/connect`.
    
    Include the token in the `Authorization` header as `Bearer <token>`.
  version: 1.0.0
  contact:
    name: Kitsu DeFi
    url: https://kitsu.fi
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.kitsu.fi
    description: Production
...
paths:
...
  /auth/connect:
    post:
      tags: [Auth]
      summary: Connect wallet
      description: Verify wallet signature and issue JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectRequest'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      description: Returns the authenticated user's profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /dens:
    get:
      tags: [Dens]
      summary: List public dens
      responses:
        '200':
          description: List of public dens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DensResponse'
    post:
      tags: [Dens]
      summary: Create den
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDenRequest'
      responses:
        '201':
          description: Den created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DenResponse'
  /dens/mine:
    get:
      tags: [Dens]
      summary: List user's dens
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User's dens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DensResponse'
  /dens/{id}:
    get:
      tags: [Dens]
      summary: Get den details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Den details with members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DenDetailResponse'
  /dens/{id}/join:
    post:
      tags: [Dens]
      summary: Join den
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinDenRequest'
      responses:
        '200':
          description: Joined den
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinDenResponse'
  /dens/{id}/leave:
    post:
      tags: [Dens]
      summary: Leave den
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Left den
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      left:
                        type: boolean
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/connect
  schemas:
...
    Success:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request body

    # Health
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        walletAddr:
          type: string
          description: TON wallet address
        username:
          type: string
          nullable: true
        tonHandle:
          type: string
          nullable: true
        xp:
          type: integer
          description: Total XP points
        level:
          type: integer
          description: Current level
        streakDays:
          type: integer
          description: Consecutive days saved
        createdAt:
          type: string
          format: date-time
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: '#/components/schemas/User'
    ConnectRequest:
      type: object
      required: [address, signature, timestamp]
      properties:
        address:
          type: string
          description: TON wallet address
        signature:
          type: string
          description: Wallet signature
        timestamp:
          type: integer
          description: Unix timestamp
    ConnectResponse:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            data:
              type: object
              properties:
                token:
                  type: string
                  description: JWT token
                user:
                  $ref: '#/components/schemas/User'
...
    Den:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        emoji:
          type: string
          nullable: true
        isPublic:
          type: boolean
        strategy:
          type: string
          enum: [steady, adventurous]
        apr:
          type: string
          description: Annual percentage rate
        totalDeposited:
          type: string
          description: Total TON deposited
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    DenMember:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
        amount:
          type: string
          description: Member's deposited amount in TON
    DenWithMembers:
      allOf:
        - $ref: '#/components/schemas/Den'
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: '#/components/schemas/DenMember'
    DenResponse:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            data:
              type: object
              properties:
                den:
                  $ref: '#/components/schemas/Den'
    DenDetailResponse:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            data:
              type: object
              properties:
                den:
                  $ref: '#/components/schemas/DenWithMembers'
    DensResponse:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            data:
              type: object
              properties:
                dens:
                  type: array
                  items:
                    $ref: '#/components/schemas/Den'
    CreateDenRequest:
      type: object
      required: [name, strategy]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        emoji:
          type: string
          maxLength: 8
        isPublic:
          type: boolean
          default: true
        strategy:
          type: string
          enum: [steady, adventurous]
    JoinDenRequest:
      type: object
      required: [amountTon]
      properties:
        amountTon:
          type: string
          pattern: '^\d+(\.\d{1,8})?$'
          description: Amount in TON to deposit
    JoinDenResponse:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            data:
              type: object
              properties:
                deposit:
                  type: object
                  properties:
                    denId:
                      type: string
                      format: uuid
                    amount:
                      type: string
                    txParams:
                      type: object
                      description: Transaction parameters for TonConnect
                      properties:
                        to:
                          type: string
                        amount:
                          type: string
                        payload:
                          type: string

    # Portfolio
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Missing or invalid authorization header
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid request body